integration_django_vue/backend/src/sepulkas/api/serializer.py
from rest_framework import serializers
from sepulkas.models import Sepulka


class SepulkaSerializer(serializers.ModelSerializer):
    class Meta:
        model = Sepulka
        fields = [
            'id',
            'coefficient',
            'slug',
        ]
    