integration_django_vue/backend/src/requirements.txt
asgiref==3.3.1
attrs==20.3.0
Django==3.1.7
django-rest-framework==0.1.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
flake8==3.8.4
flake8-bugbear==21.3.2
flake8-commas==2.0.0
flake8-isort==4.0.0
flake8-print==4.0.0
isort==5.7.0
mccabe==0.6.1
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==2.0.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
testfixtures==6.17.1
